Update structure allocation
Changed structure allocation for VIEW and CTX structures. If you have
a structure you can pass it into the init functions and it will be
used, if you don't, then the function will use malloc to get you a
structure. In both cases they are cleared to zero. But this lets you
re-use a structure if you want to.